refactor(simple-tools): extract IntentParser into intent_parser.py
